+++ poium/javascript.py
@@ -50,6 +50,20 @@ class CSSElement(object):
                     elm.value = "";""".format(css=self.css, index=self.index)
         self._execute_javascript(js)
 
+    def get_text(self, i=None):
+        """
+        JavaScript API, Only support css positioning
+        Get element text content.
+        :param i: index
+        """
+        if i is None:
+            i = self.index
+
+        logging.info("Element of the current operation: {desc}".format(desc=self.desc))
+        js = """return document.querySelectorAll("{css}")[{index}].textContent;""".format(
+            css=self.css, index=i)
+        return self._execute_javascript(js)
+
     def set_text(self, value):
         """
         JavaScript API, Only support css positioning